Finished NonIntersect Part 1
A very small bug still remains that despite two following vectors x and x+1 being incapable of intersecting they can very rarely display an intersection message even tho they aren't. Could fix by adding an if statement on printing intersections but the real fix is on the math somewhere. Part two is to calculate the area of the given solid